# api/openapi.yaml
openapi: "3.0.3"
info:
  title: "Direct P2P - Local Agent API"
  description: |
    Direct P2P 로컬 에이전트 API 명세서.
    사용자 PC에서 실행되는 로컬 에이전트와 통신하기 위한 API입니다.
    localhost(127.0.0.1)에만 바인딩되어 외부 접근이 차단됩니다.
    세션 요청 시 토큰을 발급하여 다른 프로그램이 이 세션을 가로채지 못하도록 합니다.
  version: "1.0.0"
  contact:
    name: "Direct P2P 개발팀"
servers:
  - url: "http://127.0.0.1:17432"
    description: "로컬 에이전트 (localhost 전용)"
...
paths:
...
  /session/start:
    post:
      tags:
        - "세션"
      summary: "새 세션 시작 (송신자)"
      description: |
        송신자가 새로운 P2P 세션을 시작합니다.
      operationId: "startSession"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSessionRequest"
      responses:
        "201":
          description: "세션 생성 및 초대 코드 발급 성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartSessionResponse"
        "400":
          description: "잘못된 요청 (파일 목록 누락 등)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: "이미 활성 세션이 존재함"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "에이전트 내부 오류"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    StartSessionRequest:
      type: object
      description: "세션 시작 요청 (송신자)"
      required:
        - targetUser
      properties:
        targetUser:
          type: string
          description: "P2P 연결할 상대방 퍼블릭 ip:port"

    # ──────────────────────────────────────────────
    # 4. StartSessionResponse — POST /session/start 응답
    # ──────────────────────────────────────────────
    StartSessionResponse:
      type: object
      description: "세션 시작 응답"
      required:
        - sessionId
      properties:
        sessionId:
          type: string
          description: "생성된 세션 ID"
          example: "sess_abc123"

    # ──────────────────────────────────────────────
    # 5. FileInfo — StartSessionRequest, SendDataRequest에서 재사용
    # ──────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      description: "에러 응답"
      required:
        - code
        - message
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: "에러 메시지"
          example: "세션을 찾을 수 없습니다."
        detail:
          type: string
          nullable: true
          description: "에러 상세 정보 (없으면 null)"
          example: "sessionId: sess_abc123에 해당하는 세션이 존재하지 않습니다."

    # ──────────────────────────────────────────────
    # 15. ErrorCode — ErrorResponse에서 $ref 참조
    # ──────────────────────────────────────────────
    ErrorCode:
      type: string
      description: |
        에러 코드 목록:
        - E1001: 잘못된 요청 파라미터
        - E1002: 필수 필드 누락
        - E2001: 세션을 찾을 수 없음
        - E2002: 세션이 이미 존재함 (중복 생성 시도)
        - E3001: 피어를 찾을 수 없음
        - E3002: 피어 연결 실패
        - E4001: 전송 초기화 실패
        - E4002: 전송 중 오류 발생
        - E4003: 전송이 이미 진행 중
        - E5001: NAT 탐지 실패
        - E5002: 홀펀칭 실패
        - E6001: 에이전트 내부 오류
      enum:
        - E1001
        - E1002
        - E2001
        - E2002
        - E3001
        - E3002
        - E4001
        - E4002
        - E4003
        - E5001
        - E5002
        - E6001